시큐리티 설정 패키지 변경
